requirements.txt
streamlit
ultralytics
Pillow
langchain
faiss-cpu
transformers
sentence-transformers
